Towers can be placed only on specific areas
-Mouse cursor position improved
-Towers snap properly now
-Tower UI created placed at the bottom of the screen
