+++ Assets/Scripts/Tower Defense/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bc3170ccbf9e61940b7ac38df68aacc0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
