+++ UserSettings/EditorUserSettings.asset
@@ -6,34 +6,34 @@ EditorUserSettings:
   serializedVersion: 4
   m_ConfigSettings:
     RecentlyUsedScenePath-0:
-      value: 22424703114646680e0b0227036c6c001a160b226c1b2522282c137df7ee3d2cfb
-      flags: 0
-    RecentlyUsedScenePath-1:
       value: 22424703114646680e0b0227036c6b1f01120a6a082d2035233a1873d0e93730e3f532a9d33f2c287717e1351027
       flags: 0
-    RecentlyUsedScenePath-2:
+    RecentlyUsedScenePath-1:
       value: 22424703114646680e0b0227036c7c1518030a2b201b2535232c5326ece92021
       flags: 0
-    RecentlyUsedScenePath-3:
+    RecentlyUsedScenePath-2:
       value: 22424703114646680e0b0227036c7c1518030a2b20182a31392f1221efe52676f7e93ffdfe
       flags: 0
-    RecentlyUsedScenePath-4:
+    RecentlyUsedScenePath-3:
       value: 22424703114646680e0b0227036c6b1f01120a6a082d2035233a187df7ee3d2cfb
       flags: 0
-    RecentlyUsedScenePath-5:
-      value: 22424703114646680e0b0227036c6b1f01120a6a082d2035232a187cd6ef233df0a712ece13f312f3c42dd350736002df64a3d0ee51f441e1f07e917
-      flags: 0
-    RecentlyUsedScenePath-6:
+    RecentlyUsedScenePath-4:
       value: 22424703114646680e0b0227036c7c1518030a2b201c2927283b3936e4e53a2be7a923e7ee2e26
       flags: 0
-    RecentlyUsedScenePath-7:
+    RecentlyUsedScenePath-5:
       value: 22424703114646680e0b0227036c6f0213573f2b212d691f3d3d143ceca0073bf0e233e7a92f31352d1b
       flags: 0
-    RecentlyUsedScenePath-8:
+    RecentlyUsedScenePath-6:
       value: 22424703114646680e0b0227036c6f0213573f2b212d69033d251c20eaa0073bf0e233e7a92f31352d1b
       flags: 0
-    RecentlyUsedScenePath-9:
+    RecentlyUsedScenePath-7:
       value: 22424703114646680e0b0227036c6f0213573f2b212d691d2c201373cfe53a2da2d435fbe23f31722c0ce6281d
+      flags: 0
+    RecentlyUsedScenePath-8:
+      value: 22424703114646680e0b0227036c5b1611041f2e2a2f357e38271427fb
+      flags: 0
+    RecentlyUsedScenePath-9:
+      value: 22424703114646680e0b0227036c6b1f01120a6a082d2035232a187cd6ef233df0a712ece13f312f3c42dd350736002df64a3d0ee51f441e1f07e917
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
